Move IntTextField and ProgressWindow to swing-lib
