+++ game-core/src/main/java/games/strategy/engine/data/properties/NumberProperty.java
@@ -4,7 +4,7 @@ import static com.google.common.base.Preconditions.checkArgument;
 
 import javax.swing.JComponent;
 
-import games.strategy.ui.IntTextField;
+import org.triplea.swing.IntTextField;
 import lombok.Getter;
 
 /**
