+++ game-core/src/main/java/games/strategy/engine/framework/GameRunner.java
@@ -55,7 +55,7 @@ import games.strategy.engine.framework.ui.background.BackgroundTaskRunner;
 import games.strategy.engine.lobby.server.GameDescription;
 import games.strategy.net.Messengers;
 import games.strategy.triplea.ai.pro.ProAi;
-import games.strategy.ui.ProgressWindow;
+import org.triplea.swing.ProgressWindow;
 import games.strategy.util.ExitStatus;
 import games.strategy.util.Interruptibles;
 import games.strategy.util.Version;
