+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/ClientOptions.java
@@ -16,7 +16,7 @@ import javax.swing.JTextField;
 
 import org.triplea.swing.SwingAction;
 
-import games.strategy.ui.IntTextField;
+import org.triplea.swing.IntTextField;
 
 /**
  * UI for choosing client options.
