+++ game-core/src/main/java/games/strategy/ui/ScrollableTextField.java
@@ -15,6 +15,8 @@ import javax.swing.JButton;
 import javax.swing.JPanel;
 
 import games.strategy.engine.framework.system.SystemProperties;
+import org.triplea.swing.IntTextField;
+import org.triplea.swing.IntTextFieldChangeListener;
 
 /**
  * A UI component that displays a scrollable text field for inputting integers. Four buttons are provided to change the
