+++ game-core/src/main/java/tools/map/making/MapPropertiesMaker.java
@@ -51,7 +51,7 @@ import javax.swing.SwingUtilities;
 import org.triplea.swing.SwingAction;
 
 import games.strategy.engine.data.properties.PropertiesUi;
-import games.strategy.ui.IntTextField;
+import org.triplea.swing.IntTextField;
 import games.strategy.util.Tuple;
 import lombok.extern.java.Log;
 import tools.image.FileSave;
