+++ swing-lib/src/main/java/org/triplea/swing/IntTextFieldChangeListener.java
@@ -1,4 +1,4 @@
-package games.strategy.ui;
+package org.triplea.swing;
 
 /**
  * A listener that is notified of changes to instances of {@link IntTextField}.
